+++ spec/functional/mongoid/document_spec.rb
@@ -300,35 +300,11 @@ describe Mongoid::Document do
       @person = Person.create(:title => "Test")
     end
 
-    context "finding all documents" do
-
-      it "returns an array of documents based on the selector provided" do
-        documents = Person.find(:all, :conditions => { :title => "Test"})
-        documents.first.title.should == "Test"
-      end
-    end
-
-    context "finding first document" do
-
-      it "returns the first document based on the selector provided" do
-        person = Person.find(:first, :conditions => { :title => "Test" })
-        person.title.should == "Test"
-      end
-    end
-
     context "finding by id" do
 
       it "finds the document by the supplied id" do
         person = Person.find(@person.id)
         person.id.should == @person.id
-      end
-    end
-
-    context "limiting result fields" do
-
-      it "adds the type field to the options" do
-        people = Person.all(:fields => [ :title ])
-        people.first.title.should == "Test"
       end
     end
   end
@@ -464,7 +440,7 @@ describe Mongoid::Document do
       end
 
       it "saves the entire graph up from the has_one" do
-        person = Person.first(:conditions => { :title => "Sir" })
+        person = Person.where(:title => "Sir").first
         person.should == @person
       end
     end
@@ -476,7 +452,7 @@ describe Mongoid::Document do
       end
 
       it "saves the entire graph up from the embeds_many" do
-        person = Person.first(:conditions => { :title => "Sir" })
+        person = Person.where(:title => "Sir").first
         person.should == @person
       end
     end